Move all logic from Wallet component into wallet composable
